src/auth_system/requirements.txt
# Authentication app dependencies
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django>=4.2.0